Rename accountReducers to accountsReducers.
